+++ app/(pola)/user/profile/[nickname]/page.tsx
@@ -4,9 +4,7 @@ import { useParams } from "next/navigation";
 import UserTierSection from "@/app/_components/sections/user-tier/UserTierSection";
 import UserArchivmentSection from "@/app/_components/sections/user-archivment/UserArchivmentSection";
 import UserHelpsSection from "@/app/(pola)/user/profile/[nickname]/_components/user-helps/UserHelpsSection";
-import ProfileMenuSection from "./_components/sections/ProfileMenuSection";
-import { useEffect } from "react";
-import { useAuthStore } from "@/lib/stores/authStore";
+import ProfileMenuSection from "./_components/sections/profile-menu/ProfileMenuSection";
 import { useApiQuery } from "@/lib/hooks/useApi";
 import { UserProfileResponseDto } from "@/backend/users/user/applications/dtos/UserDtos";
 import UserInfoSection from "@/app/_components/commons/common-sections/user-info/UserInfoSection";
